my_project/requirements.txt
APScheduler==3.8.1
Django==3.2.9
django-allauth==0.47.0
django-ckeditor==6.2.0
django-constance==2.8.0
django-js-asset==1.2.2
django-picklefield==3.0.1
django-uploader==0.2.1
django-redis==5.2.0
vonage==3.2.2
sorl-thumbnail==12.7.0
Pillow==9.0.0
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
idna==3.3
oauthlib==3.1.1
celery==5.2.3
pytest==7.2.0
channels==3.0.5
channels-redis==3.4.1
